generated_scripts/script_20251103_012145/requirements.txt
loguru
psycopg2-binary
pydantic
python-dotenv